% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{create_api_package}
\alias{create_api_package}
\title{Create API package}
\usage{
create_api_package(path, site, base_url, ...)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{site}{Name of the site/app/platform. This is how it will be referenced
in function documentation.}

\item{base_url}{Base URL to the site/app/platform of interest.}

\item{...}{Other arguments passed on to \link[usethis]{create_package}.}
}
\value{
Creates or uses existing directory and builds the basic
  infastructure for an API client R package.
}
\description{
Create API package
}
\details{
In addition to \link[usethis]{create_package} defaults, this
  function provides useful starter functions for creating an R client for
  interacting with a web API. The functions are saved in two files in the
  path's R directory. api.R contains functions for sending/receiving HTTP
  requests. utils.R contains functions for setting environment variables and
  parsing response objects.
}
\examples{
\dontrun{
## create API client package for Instagram API
create_api_package("example/useapitest",
  "Use API Test Site", "https://api.useapitest.com/v1")
}
}
